+++ client/components/social-buttons/index.ts
@@ -1,0 +1,5 @@
+export { default as MagicLoginButton } from './magic-login';
+export { default as QrCodeLoginButton } from './qr-code';
+export { default as GoogleSocialButton } from './google';
+export { default as AppleLoginButton } from './apple';
+export { default as GithubSocialButton } from './github';
