+++ client/components/social-buttons/qr-code.tsx
@@ -1,0 +1,77 @@
+import { recordTracksEvent } from '@automattic/calypso-analytics';
+import { Button } from '@wordpress/components';
+import clsx from 'clsx';
+import { useTranslate } from 'i18n-calypso';
+import JetpackLogo from 'calypso/components/jetpack-logo';
+import { login } from 'calypso/lib/paths';
+import { useSelector, useDispatch } from 'calypso/state';
+import { resetMagicLoginRequestForm } from 'calypso/state/login/magic-login/actions';
+import { isFormDisabled } from 'calypso/state/login/selectors';
+import { getCurrentOAuth2Client } from 'calypso/state/oauth2-clients/ui/selectors';
+import getCurrentLocaleSlug from 'calypso/state/selectors/get-current-locale-slug';
+import getCurrentQueryArguments from 'calypso/state/selectors/get-current-query-arguments';
+import getIsWooPasswordless from 'calypso/state/selectors/get-is-woo-passwordless';
+
+type QrCodeLoginButtonProps = {
+	twoFactorAuthType: string;
+};
+
+const QrCodeLoginButton = ( { twoFactorAuthType }: QrCodeLoginButtonProps ) => {
+	const dispatch = useDispatch();
+	const translate = useTranslate();
+	const { isDisabled, query, locale, oauth2Client, isWooPasswordless } = useSelector(
+		( select ) => {
+			return {
+				isDisabled: isFormDisabled( select ),
+				query: getCurrentQueryArguments( select ),
+				oauth2Client: getCurrentOAuth2Client( select ) as { id: string },
+				locale: getCurrentLocaleSlug( select ),
+				isWooPasswordless: getIsWooPasswordless( select ),
+			};
+		}
+	);
+	const isJetpackWooCommerceFlow = 'woocommerce-onboarding' === query?.from;
+
+	if ( twoFactorAuthType ) {
+		return null;
+	}
+	// Is not supported for any oauth 2 client.
+	// n.b this seems to work for woo.com so it's not clear why the above comment is here
+	if ( oauth2Client && ! isWooPasswordless ) {
+		return null;
+	}
+
+	if ( isJetpackWooCommerceFlow ) {
+		return null;
+	}
+
+	const handleClick = () => {
+		recordTracksEvent( 'calypso_login_magic_login_request_click', {
+			origin: 'login-links',
+		} );
+
+		dispatch( resetMagicLoginRequestForm() );
+	};
+
+	const loginUrl = login( {
+		locale: locale,
+		twoFactorAuthType: 'qr',
+		redirectTo: query?.redirect_to as string,
+		signupUrl: query?.signup_url as string,
+	} );
+
+	return (
+		<Button
+			className={ clsx( 'social-buttons__button button', { disabled: isDisabled } ) }
+			href={ loginUrl }
+			onClick={ handleClick }
+			data-e2e-link="magic-login-link"
+			key="lost-password-link"
+		>
+			<JetpackLogo monochrome={ isDisabled } size={ 20 } />
+			<span className="social-buttons__service-name">{ translate( 'Login via app' ) }</span>
+		</Button>
+	);
+};
+
+export default QrCodeLoginButton;
